broker/requirements.txt
kafka-python
cisu-python
dataclasses_json
dataclasses
elasticsearch